Separate Network construction and system assembler
